+++ app/(landing)/layout.tsx
@@ -1,11 +1,5 @@
 import { Footer } from "@/components/footer";
 import { Header } from "@/components/header";
-import type { Metadata } from "next";
-
-export const metadata: Metadata = {
-  title: "Articuria",
-  description: "Practice public speaking at home",
-};
 
 export default function RootLayout({
   children,
